Refactor profile_page.
