+++ gui_code/gui.py
@@ -1,14 +1,11 @@
 from tkinter import Frame, Label, Button, INSIDE, BOTH, RIGHT, LEFT, Tk
-import tkinter
-import matplotlib.pyplot as plt
-from sensor_value_scale import SensorValueScale
-from profileInformation import ProfileInformation
-from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
 from PIL import Image
 from PIL import ImageTk
 from matplotlib.animation import FuncAnimation
 from matplotlib import style
 import time
+
+from profile_page import ProfilePage
 
 style.use("ggplot")
 
@@ -31,25 +28,6 @@ class FarmBeatsApp:
         self.option_frame_setup()
         self.profile_frame = Frame(self.main, bg="white")
 
-        self.sensor_frame = None
-        self.curr_sensor_value_label = None
-        self.curr_actuator_value_label = None
-        self.actuator_frame = None
-        self.suggestion_frame = None
-
-        self.water_level_frame = None
-        self.is_water = False
-
-        # for graph display
-        self.graph_frame = None
-        self.current_profile = None
-        self.time_since_update = time.time()
-        self.fig = None
-        self.plot = None
-        self.data_plot = None
-        self.canvas = None
-        self.widget = None
-        self.animation = None
 
     def label_frame_setup(self):
         self.label = Label(self.label_frame, text="IoT FarmBeats", width=60)
@@ -84,8 +62,7 @@ class FarmBeatsApp:
 
         tempButton.image = temp_img
         tempButton.grid(
-            row=0, column=0, sticky="news",
-            pady=PADDING, padx=PADDING
+            row=0, column=0, sticky="news", pady=PADDING, padx=PADDING
         )
 
         humidity_img = self.get_menu_button(
@@ -101,8 +78,7 @@ class FarmBeatsApp:
 
         humidityButton.image = humidity_img
         humidityButton.grid(
-            row=0, column=1, sticky="news",
-            pady=PADDING, padx=PADDING
+            row=0, column=1, sticky="news", pady=PADDING, padx=PADDING
         )
 
         brightness_img = self.get_menu_button(
@@ -133,8 +109,7 @@ class FarmBeatsApp:
 
         waterButton.image = water_img
         waterButton.grid(
-            row=1, column=0, sticky="news",
-            pady=PADDING, padx=PADDING
+            row=1, column=0, sticky="news", pady=PADDING, padx=PADDING
         )
 
         aiCamera_img = self.get_menu_button(
@@ -150,8 +125,7 @@ class FarmBeatsApp:
 
         aiCameraButton.image = aiCamera_img
         aiCameraButton.grid(
-            row=1, column=1, sticky="news",
-            pady=PADDING, padx=PADDING
+            row=1, column=1, sticky="news", pady=PADDING, padx=PADDING
         )
 
         system_img = self.get_menu_button(
@@ -167,331 +141,27 @@ class FarmBeatsApp:
 
         sysVisualButton.image = system_img
         sysVisualButton.grid(
-            row=1, column=2, sticky="news",
-            pady=PADDING, padx=PADDING
+            row=1, column=2, sticky="news", pady=PADDING, padx=PADDING
         )
 
-        self.option_frame.pack(expand=True, fill=BOTH, pady=15, padx=15)
-
-    def general_actuation_setup(self):
-
-        self.actuator_frame = Frame(self.profile_frame, bg=BACKGROUND)
-        self.actuator_frame.grid_columnconfigure(0, weight=1)
-        for n in range(3):
-            self.actuator_frame.grid_rowconfigure(n, weight=1)
-
-        actuatorTitle = Label(
-            self.actuator_frame,
-            text=self.profile.actuator_frame_title
-        )
-
-        actuatorTitle.config(background=BACKGROUND, font=("Courier", 15))
-
-        actuatorTitle.grid(row=0, column=0, sticky="news", padx=PADDING)
-
-        mode_switch_frame = Frame(self.actuator_frame, bg=BACKGROUND)
-        manual_mode = Label(mode_switch_frame, text="Manual")
-        manual_mode.config(background="#CEE5DB", font=("Courier", 15))
-        manual_mode.pack(side=RIGHT)
-        automatic_mode = Button(mode_switch_frame, text="Automatic")
-        automatic_mode.config(background=BACKGROUND, font=("Courier", 15))
-        automatic_mode.pack(side=LEFT)
-
-        mode_switch_frame.grid(row=1, column=0, sticky="news", padx=PADDING)
-
-        self.curr_actuator_value_label = Label(
-            self.actuator_frame, text=self.profile.actuator_value_description
-        )
-
-        self.curr_actuator_value_label.config(
-            background=BACKGROUND, font=("Courier", 15)
-        )
-
-        self.curr_actuator_value_label.grid(
-            row=2, column=0, sticky="news", padx=PADDING
-        )
-
-        if self.is_water:
-            self.actuator_frame.grid(
-                row=0,
-                column=1,
-                sticky="news",
-                pady=PADDING,
-                padx=PADDING,
-            )
-            self.water_level_frame_setup()
-        else:
-            self.actuator_frame.grid(
-                row=0, column=1, sticky="news",
-                pady=PADDING, padx=PADDING, rowspan=2
-            )
-
-    def water_level_frame_setup(self):
-
-        water_level_value = 50  # dummy data for now
-
-        self.water_level_frame = Frame(self.profile_frame, bg=BACKGROUND)
-        self.water_level_frame.grid_columnconfigure(0, weight=1)
-        self.water_level_frame.grid_rowconfigure(0, weight=1)
-        self.water_level_frame.grid_rowconfigure(1, weight=3)
-        self.water_level_frame.grid_rowconfigure(2, weight=1)
-
-        water_level_title = Label(
-            self.water_level_frame,
-            text="Water Level: " + str(water_level_value) + "%"
-        )
-
-        water_level_title.config(background=BACKGROUND, font=("Courier", 15))
-
-        water_level_title.grid(row=0, column=0, sticky="news", padx=PADDING)
-
-        scale_frame = Frame(self.water_level_frame)
-
-        h = 200
-        y_offset = 20
-        x_offset = 10
-
-        scale_canvas = tkinter.Canvas(
-            scale_frame,
-            height=h + y_offset,
-            width=80 + (2 * x_offset),
-        )
-
-        length = (water_level_value / 100) * h
-
-        scale_canvas.create_rectangle(
-            x_offset, y_offset, 50 + x_offset, h + y_offset, width=0
-        )
-        scale_canvas.create_rectangle(
-            x_offset,
-            h + y_offset,
-            50 + x_offset,
-            h - length + y_offset,
-            fill="#B7DEF2",
-            width=0,
-        )
-        scale_canvas.create_line(
-            x_offset, h + y_offset, x_offset,
-            y_offset, fill="#000000", width=3
-        )
-        scale_canvas.create_line(
-            49 + x_offset,
-            h + y_offset,
-            49 + x_offset,
-            y_offset,
-            fill="#000000",
-            width=3,
-        )
-        scale_canvas.create_line(
-            50 + x_offset, h + y_offset, x_offset,
-            h + y_offset, fill="#000000", width=3
-        )
-        scale_canvas.create_text(
-            50 + (2 * x_offset) + 10,
-            y_offset,
-            text="100%",
-            fill="black",
-            font=("Courier"),
-        )
-        scale_canvas.create_text(
-            50 + (2 * x_offset) + 10,
-            h // 2 + y_offset - 5,
-            text="50%",
-            fill="black",
-            font=("Courier"),
-        )
-        scale_canvas.create_text(
-            50 + (2 * x_offset) + 10,
-            h + y_offset - 5,
-            text="0%",
-            fill="black",
-            font=("Courier"),
-        )
-
-        scale_canvas.pack()
-        scale_frame.grid(
-            row=1,
-            column=0,
-            sticky="news",
-            padx=PADDING,
-        )
-
-        water_level_title.config(
-            background=BACKGROUND, font=("Courier", 15)
-        )
-
-        water_level_title.grid(row=0, column=0, sticky="news", padx=PADDING)
-
-        self.water_level_frame.grid(
-            row=1,
-            column=1,
-            sticky="news",
-            pady=PADDING,
-            padx=PADDING,
-        )
-
-    def profile_setup(self, profile_name):
-        self.profile = ProfileInformation(profile_name)
-        self.label.config(text=self.profile.title)
-        img = Image.open("assets/homeIcon.png")
-        home_icon = ImageTk.PhotoImage(img)
-
-        home_button = Button(
-            self.label_frame, image=home_icon, borderwidth=0
-        )
-        home_button.image = home_icon
-
-        home_button[
-            "command"
-        ] = lambda idx="Home", \
-            binst=home_button: self.home_button_action(binst)
-
-        home_button.pack()
-        home_button.place(bordermode=INSIDE, x=5, y=5)
-
-        for n in range(2):
-            self.profile_frame.grid_columnconfigure(n, weight=1)
-        for n in range(3):
-            self.profile_frame.grid_rowconfigure(n, weight=1)
-
-        self.sensor_frame = Frame(self.profile_frame, bg=BACKGROUND)
-
-        sensor_title = Label(
-            self.sensor_frame, text=self.profile.sensor_frame_title
-        )
-
-        sensor_title.config(background=BACKGROUND, font=("Courier", 15))
-        sensor_title.pack()
-
-        self.curr_sensor_value_label = Label(
-            self.sensor_frame, text=self.profile.sensor_value_description
-        )
-        self.curr_sensor_value_label.config(
-            background=BACKGROUND, font=("Courier", 15)
-        )
-        self.curr_sensor_value_label.pack()
-
-        # scale_frame set up
-
-        SensorValueScale(self.profile, self.sensor_frame)
-
-        self.graph_display()
-
-        self.sensor_frame.grid(
-            row=0,
-            column=0,
-            sticky="news",
-            pady=PADDING,
-            padx=PADDING,
-            rowspan=3,
-        )
-
-        # actuator_frame set up
-
-        self.general_actuation_setup()
-
-        # suggestion_frame set up
-
-        self.suggestion_frame = Frame(self.profile_frame, bg=BACKGROUND)
-        suggestion_label = Label(self.suggestion_frame, text="Suggestion")
-        suggestion_label.config(background=BACKGROUND, font=("Courier", 15))
-        suggestion_label.pack()
-        message_frame = Frame(self.suggestion_frame, bg="#FFFFFF", height=400)
-        msg = Label(message_frame, text=self.profile.suggestion)
-        msg.pack()
-        message_frame.pack()
-
-        self.suggestion_frame.grid(
-            row=2, column=1, sticky="news", pady=PADDING, padx=PADDING
-        )
-
-        # Display on profile_frame
-
-        self.profile_frame.pack(fill=BOTH, expand=True, pady=15, padx=15)
-
-    def graph_display(self):
-        self.graph_frame = Frame(self.sensor_frame, bg=BACKGROUND)
-
-        y_label = self.profile.title + " (" + self.profile.unit + ")"
-        xar = self.profile.time_list[-100:]
-        yar = self.profile.val_list[-100:]
-        self.fig = plt.figure(figsize=(5, 4), dpi=100, tight_layout=True)
-
-        self.axs = self.fig.add_subplot(111)
-        self.axs.plot(xar, yar)
-
-        # loc = dates.AutoDateLocator(minticks=4, maxticks=4)
-        # self.fig.subplots().xaxis.set_major_locator(loc)
-        # fmt = dates.AutoDateFormatter(loc)
-        # self.fig.subplots().xaxis.set_major_formatter(fmt)
-
-        self.canvas = FigureCanvasTkAgg(
-            self.fig, self.graph_frame
-        )
-
-        self.canvas.draw()
-        self.canvas.get_tk_widget().pack(pady=15, padx=15)
-        self.animation = FuncAnimation(
-            self.fig, self.animate, interval=TIME_INTERVAL
-        )
-
-        self.axs.set(
-            xlabel="Time (ms)", ylabel=y_label,
-            title=self.profile.graph_title
-        )
-
-        self.graph_frame.pack()
-
-    def animate(self, i):
-        if time.time() - self.time_since_update >= TIME_INTERVAL:
-            self.profile.update_from_db(self.profile.title)
-            self.curr_sensor_value_label.config(
-                text=self.profile.sensor_value_description
-            )
-            self.curr_actuator_value_label.config(
-                text=self.profile.actuator_value_description
-            )
-            self.time_since_update = time.time()
-
-        if len(self.profile.time_list) < 100:
-            xar = self.profile.time_list
-            yar = self.profile.val_list
-        else:
-            xar = self.profile.time_list[-100:]
-            yar = self.profile.val_list[-100:]
-
-        self.axs.clear()
-        self.axs.plot(xar, yar)
-        self.axs.set(
-            xlabel="Time (ms)",
-            ylabel=self.profile.title + " (" + self.profile.unit + ")",
-            title=self.profile.graph_title,
-        )
-
-    def home_button_action(self, binst):
-        self.is_water = False
-        self.animation.event_source.stop()
-        self.profile_frame.pack_forget()
-        self.option_frame.pack(expand=True, fill=BOTH, pady=15, padx=15)
-        binst.destroy()
-        self.label.config(text="IoT FarmBeats")
+        self.option_frame.pack(expand=True, fill=BOTH, pady=15, padx=15)  
 
     def temp_button_action(self):
         self.option_frame.pack_forget()
-        self.profile_setup("Temperature")
+        ProfilePage("Temperature", self.profile_frame, self.label_frame, self.label, self.option_frame)
 
     def humidity_button_action(self):
         self.option_frame.pack_forget()
-        self.profile_setup("Humidity")
+        ProfilePage("Humidity", self.profile_frame, self.label_frame, self.label, self.option_frame)
 
     def brightness_button_action(self):
         self.option_frame.pack_forget()
-        self.profile_setup("Brightness")
+        ProfilePage("Brightness", self.profile_frame, self.label_frame, self.label, self.option_frame)
 
     def water_button_action(self):
         self.is_water = True
         self.option_frame.pack_forget()
-        self.profile_setup("Water Level")
+        ProfilePage("Water Level", self.profile_frame, self.label_frame, self.label, self.option_frame)
 
     def ai_camera_button_action(self):
         self.label.config(text="AI Camera Button Clicked")
